requirements.txt
pandas
torch
torchvision
Pillow
requests
tqdm
scikit-learn